changed count_sort
Tweaked count_sort method slightly on the encodefiles class. Previously would scan twice (once for the file and once for the list of byes). This doubled the running time of the method. Now it just scans the file once and updates the frequency table as it goes through the file